feat(graphql): add BigInt scalar type for safe integer handling

+++ src/app/api/graphql/category/index.ts
@@ -24,7 +24,7 @@ export const typeDefs = `#graphql
 
     totalCount: Int
 
-    totalAmountCents: Int
+    totalAmountCents: BigInt
     averageAmountCents: Float
   }
